AV 2 Concuída!

+++ AV 2/main.py
@@ -9,40 +9,52 @@ connection = mysql.connector.connect(
 
 cursor = connection.cursor()
 
+numberDigited = input(
+    'O que você deseja fazer? \n Digite \n "1" Para CRIAR um cliente \n "2" Para VER um cliente que já foi registrado \n "3" Para ATUALIZAR um cliente \n "4" Para DELETAR um cliente \n')
+
+if numberDigited != "1" or "2" or "3" or "4":
+    print(' ----------------------------------------------------------')
+    print('   O que você digitou não é o esperado, tente novamente!   ')
+    print(' ----------------------------------------------------------')
+
 # CREATE
 
-name = 'Erick Veríssimo'
-cpf = '12345678910'
-phone = 40028922
+if numberDigited == '1':
+    name = 'Erick Veríssimo'
+    cpf = '12345678910'
+    phone = 40028922
 
-command = f'INSERT INTO clients(name, cpf, phone) VALUES ("{name}", {cpf}, {phone})'
+    command = f'INSERT INTO clients(name, cpf, phone) VALUES ("{name}", {cpf}, {phone})'
 
-cursor.execute(command)
+    cursor.execute(command)
 
-connection.commit()
+    connection.commit()
 
 # READ
 
-command = f'SELECT * FROM clients;'
-cursor.execute(command)
-result = cursor.fetchall()
-## print(result)
+if numberDigited == '2':
+    command = f'SELECT * FROM clients;'
+    cursor.execute(command)
+    result = cursor.fetchall()
+    print(result)
 
 # UPDATE
 
-cpf = '12345678910'
-newCpf = '09876543211'
+if numberDigited == '3':
+    cpf = '12345678910'
+    newCpf = '09876543211'
 
-command = f'UPDATE clients SET cpf = {newCpf} WHERE cpf = "{cpf}"'
-cursor.execute(command)
-connection.commit()
+    command = f'UPDATE clients SET cpf = {newCpf} WHERE cpf = "{cpf}"'
+    cursor.execute(command)
+    connection.commit()
 
 
 # DELETE
 
-command = f'DELETE FROM clients WHERE name = "Erick Veríssimo"'
-cursor.execute(command)
-connection.commit()
+if numberDigited == '4':
+    command = f'DELETE FROM clients WHERE name = "Erick Veríssimo"'
+    cursor.execute(command)
+    connection.commit()
 
 cursor.close()
 connection.close()
